refactor: decompose top 3 largest files (DetailsExpressiveScreen, ReaderConfigSheet, EpubReaderFragment)
- Extract 27 composables from DetailsExpressiveScreen.kt (1705→270 lines)
  into ActionDockComponents, HeroSectionComponents, ProgressComponents,
  DetailsContentComponents, DetailsCommonComponents
- Extract ReadModeSection and DoublePageConfigSection from ReaderConfigSheet.kt
  into ReaderModeComponents.kt (1849→1610 lines)
- Extract data classes and custom views from EpubReaderFragment.kt
  into EpubDataModels.kt and EpubCustomViews.kt (1400→1245 lines)
