+++ app/src/main/kotlin/org/koitharu/kotatsu/reader/ui/epub/EpubReaderFragment.kt
@@ -664,7 +664,7 @@ class EpubReaderFragment : BaseReaderFragment<FragmentReaderEpubBinding>() {
 		if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) setJustificationMode(1)
 	}
 
-	private fun createTextView(parent: ViewGroup, paged: Boolean): TextView = EpubSelectableTextView(parent.context).apply {
+	private fun createTextView(parent: ViewGroup, paged: Boolean): TextView = EpubSelectableTextView(parent.context, highlightColor).apply {
 		layoutParams = if (paged) ViewGroup.LayoutParams(-1, -1) else ViewGroup.LayoutParams(-1, -2)
 		applyTextStyle(this, paged)
 		setTextIsSelectable(true)
@@ -1208,161 +1208,6 @@ class EpubReaderFragment : BaseReaderFragment<FragmentReaderEpubBinding>() {
 		}
 	}
 
-	private class TextHolder(val text: TextView) : RecyclerView.ViewHolder(text)
-	private class HighlightColorSpan(var color: Int) : CharacterStyle(), UpdateAppearance {
-		override fun updateDrawState(tp: TextPaint) {
-			tp.bgColor = color
-		}
-	}
-
-	private class EpubSelectableTextView(context: Context) : AppCompatTextView(context) {
-		private val selectionBackgroundColor = highlightColor
-		private var selectionBackgroundSpan: SelectionBackgroundSpan? = null
-		private val handlePaint = Paint(Paint.ANTI_ALIAS_FLAG)
-		private val handleRadius = 9f * resources.displayMetrics.density
-		private val handleStemWidth = 2f * resources.displayMetrics.density
-
-		init {
-			handlePaint.color = ColorUtils.setAlphaComponent(selectionBackgroundColor, 255)
-			setHighlightColor(Color.TRANSPARENT)
-		}
-
-		override fun onAttachedToWindow() {
-			super.onAttachedToWindow()
-			if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-				textSelectHandleLeft?.mutate()?.apply { alpha = 0 }?.let(::setTextSelectHandleLeft)
-				textSelectHandleRight?.mutate()?.apply { alpha = 0 }?.let(::setTextSelectHandleRight)
-				textSelectHandle?.mutate()?.apply { alpha = 0 }?.let(::setTextSelectHandle)
-			}
-		}
-
-		override fun onSelectionChanged(selStart: Int, selEnd: Int) {
-			super.onSelectionChanged(selStart, selEnd)
-			val spannable = text as? Spannable ?: return
-			selectionBackgroundSpan?.let(spannable::removeSpan)
-			selectionBackgroundSpan = null
-			if (selStart < 0 || selEnd < 0 || selStart == selEnd) return
-			val span = SelectionBackgroundSpan(this, selectionBackgroundColor, selStart, selEnd)
-			selectionBackgroundSpan = span
-			spannable.setSpan(
-				span,
-				minOf(selStart, selEnd),
-				maxOf(selStart, selEnd),
-				Spanned.SPAN_EXCLUSIVE_EXCLUSIVE,
-			)
-		}
-
-		override fun onDraw(canvas: Canvas) {
-			super.onDraw(canvas)
-			if (Build.VERSION.SDK_INT < Build.VERSION_CODES.Q || !hasSelection() || !isFocused) return
-			val start = minOf(selectionStart, selectionEnd)
-			val end = maxOf(selectionStart, selectionEnd)
-			drawHandle(canvas, start)
-			drawHandle(canvas, end)
-		}
-
-		private fun drawHandle(canvas: Canvas, offset: Int) {
-			val textLayout = layout ?: return
-			val line = textLayout.getLineForOffset(offset)
-			val x = compoundPaddingLeft - scrollX + correctedHorizontal(offset, line)
-			val metrics = paint.fontMetrics
-			val y = extendedPaddingTop - scrollY + textLayout.getLineBaseline(line) + metrics.descent
-			canvas.drawRect(x - handleStemWidth / 2f, y, x + handleStemWidth / 2f, y + handleRadius, handlePaint)
-			canvas.drawCircle(x, y + handleRadius, handleRadius, handlePaint)
-		}
-
-		fun correctedHorizontal(offset: Int, line: Int): Float {
-			val textLayout = layout ?: return 0f
-			val horizontal = textLayout.getPrimaryHorizontal(offset)
-			if (Build.VERSION.SDK_INT < 26 || justificationMode != Layout.JUSTIFICATION_MODE_INTER_WORD) return horizontal
-			val lineStart = textLayout.getLineStart(line)
-			val lineEnd = textLayout.getLineEnd(line)
-			if (lineEnd >= text.length || text[lineEnd - 1] == '\n') return horizontal
-			var visibleEnd = lineEnd
-			while (visibleEnd > lineStart && text[visibleEnd - 1].isWhitespace()) visibleEnd--
-			val spaces = (lineStart until visibleEnd).count { text[it] == ' ' }
-			if (spaces == 0) return horizontal
-			val extraPerSpace = (textLayout.width - paint.measureText(text, lineStart, visibleEnd)) / spaces
-			val spacesBefore = (lineStart until offset.coerceAtMost(visibleEnd)).count { text[it] == ' ' }
-			return horizontal + textLayout.getParagraphDirection(line) * extraPerSpace * spacesBefore
-		}
-
-		fun selectionEndHorizontal(offset: Int, line: Int): Float {
-			val textLayout = layout ?: return 0f
-			if (offset < textLayout.getLineEnd(line)) return correctedHorizontal(offset, line)
-			return if (textLayout.getParagraphDirection(line) == Layout.DIR_RIGHT_TO_LEFT) {
-				textLayout.getLineLeft(line)
-			} else {
-				textLayout.getLineRight(line)
-			}
-		}
-	}
-
-	private class SelectionBackgroundSpan(
-		private val textView: EpubSelectableTextView,
-		color: Int,
-		start: Int,
-		end: Int,
-	) : LineBackgroundSpan, NoCopySpan {
-		private val selectionStart = minOf(start, end)
-		private val selectionEnd = maxOf(start, end)
-		private val backgroundPaint = Paint().apply { this.color = color }
-
-		override fun drawBackground(
-			canvas: Canvas,
-			paint: Paint,
-			left: Int,
-			right: Int,
-			top: Int,
-			baseline: Int,
-			bottom: Int,
-			text: CharSequence,
-			start: Int,
-			end: Int,
-			lineNumber: Int,
-		) {
-			val rangeStart = maxOf(start, selectionStart)
-			val rangeEnd = minOf(end, selectionEnd)
-			if (rangeStart >= rangeEnd) return
-			val x1 = textView.correctedHorizontal(rangeStart, lineNumber)
-			val x2 = textView.selectionEndHorizontal(rangeEnd, lineNumber)
-			val metrics = paint.fontMetrics
-			canvas.drawRect(
-				minOf(x1, x2),
-				baseline + metrics.ascent,
-				maxOf(x1, x2),
-				baseline + metrics.descent,
-				backgroundPaint,
-			)
-		}
-	}
-
-	private class HighlightMarker(val bookmarkId: Long) : CharacterStyle() {
-		override fun updateDrawState(tp: TextPaint) = Unit
-	}
-	private class NativeChapter(
-		val id: Long,
-		val title: String,
-		val file: File,
-		val entryName: String,
-	) {
-		@Volatile
-		var content: Spanned? = null
-		val text: Spanned get() = content ?: EMPTY_CHAPTER_TEXT
-	}
-	private data class PreparedBook(val chapters: List<NativeChapter>, val archives: Map<File, ZipFile>)
-	private data class NativePage(val chapter: Int, val start: Int, val end: Int)
-	private data class SearchResult(val chapter: Int, val title: String, val snippet: String, val offset: Int)
-	private data class Locator(val chapter: Int, val offset: Int)
-	private data class TextLocation(val chapter: Int, val baseOffset: Int)
-	private data class SelectedText(val chapter: Int, val start: Int, val end: Int, val text: String)
-	private data class DictionaryEntry(val phonetic: String, val meanings: List<DictionaryMeaning>)
-	private data class DictionaryMeaning(
-		val partOfSpeech: String,
-		val definitions: List<DictionaryDefinition>,
-		val synonyms: List<String>,
-	)
-	private data class DictionaryDefinition(val text: String, val example: String?)
 	private fun Locator.clamped(): Locator {
 		if (chapters.isEmpty()) return Locator(0, 0)
 		val c = chapter.coerceIn(chapters.indices)
